requirements.txt
netaddr
ujson
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
elasticsearch==6.3.1
elasticsearch-dsl==6.3.1
kombu==4.5.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
six==1.12.0
sqlalchemy==1.3.2
urllib3==1.24.1
vine==1.3.0
mysql-connector-python
MySQL-python
